Refactor design tokens
