+++ src/00_design_tokens/005_iconography/Icon/_Icon.scss
