+++ src/00_design_tokens/006_typography/Typography/Typography.scss
@@ -3,6 +3,8 @@
  *
  */
 
+// @todo create typography mixin (https://scarab.style/core/design-tokens/#font-sizes-line-heights)
+
 // Import from Google Fonts
 @import url("https://fonts.googleapis.com/css2?family=Arimo:wght@400;700&family=Roboto+Slab:wght@400;700&display=swap");
