+++ src/01_atoms/ExLink/ExLink.js
@@ -9,7 +9,7 @@
 
 import React from "react"
 import PropTypes from "prop-types"
-import Icon from "../../00_design_tokens/Icon/Icon"
+import Icon from "../../00_design_tokens/005_iconography/Icon/Icon"
 
 const ExLink = props => {
     return (
