--- src/10_styles/00_base-styles/_theme.scss
@@ -1,57 +1,0 @@
-/*
- * Theme Variables - Design Choices using Design Tokens
- * 
- * Design Tokens can also be stored as a JSON file to be shared accross multiple platforms, for more information see:
- * Tokens in Design Systems by Nathan Curtis https://medium.com/eightshapes-llc/tokens-in-design-systems-25dd82d58421
- * What Are Design Tokens? by Robin Rendle https://css-tricks.com/what-are-design-tokens
- * Design Tokens with Figma by Pavel Laptev https://blog.prototypr.io/design-tokens-with-figma-aef25c42430f
- */
-
-/* Design Choices
- ****************/
-
-// // Set Site Theme Fonts
-// $theme-font-primary: $font--arimo;
-// $theme-font-secondry: $font--slab;
-
-// // Set Site Theme Color Palette
-// $theme-color-primary: $color-black;
-// $theme-color-primary--lighter: $color-black;
-// $theme-color-primary--darker: $color-black;
-
-// $theme-color-seconday: $color-indigo;
-// $theme-color-seconday--lighter: $color-indigo--lighter;
-// $theme-color-seconday--darker: $color-indigo--darker;
-
-// $theme-color-accent: $color-indigo;
-// $theme-color-accent--lighter: $color-indigo--lighter;
-// $theme-color-accent--darker: $color-indigo--darker;
-
-// $theme-color-background: $color-white;
-
-// // Set Site Background Color
-// $site-background-color: $theme-color-background;
-
-// // Set Max Width for Site
-// $site-max-width: $breakpoint-desktop;
-
-// // Set Site Heading Styles
-// $heading-font: $theme-font-primary;
-// $heading-weight: $font-weight--bold;
-// $heading-color: $theme-color-primary;
-
-// // Set Site Paragraph Styles
-// $paragraph-font: $theme-font-secondry;
-// $paragraph-weight: $font-weight--normal;
-// $paragraph-color: $theme-color-primary;
-
-// // Set Site Link Styles
-// $link-font: $theme-font-primary;
-// $link-weight: $font-weight--bold;
-// $link-color: $theme-color-primary;
-
-// // Set Site Button Styles
-// $button-font: $theme-font-primary;
-// $button-color: $color-white;
-// $button-background: $theme-color-accent;
-// $button-background--hover: $theme-color-accent--darker;
