--- src/10_styles/00_base-styles/_tokens.scss
@@ -1,83 +1,0 @@
-/*
- * Design Token Variables
- *
- * Includes Typography, Font Sizing, Color Palette, Gutter Spacing, and Breakpoints
- * 
- * These Tokens can also be stored as a JSON file to be shared accross multiple platforms, for more information see:
- * Tokens in Design Systems by Nathan Curtis https://medium.com/eightshapes-llc/tokens-in-design-systems-25dd82d58421
- * What Are Design Tokens? by Robin Rendle https://css-tricks.com/what-are-design-tokens
- * Design Tokens with Figma by Pavel Laptev https://blog.prototypr.io/design-tokens-with-figma-aef25c42430f
- */
-
-/* Typography
- ************/
-
-// // Import from Google Fonts
-// @import url("https://fonts.googleapis.com/css2?family=Arimo:wght@400;700&family=Roboto+Slab:wght@400;700&display=swap");
-
-// // Font Families
-// $font--arimo: "Arimo", sans-serif;
-// $font--slab: "Roboto Slab", serif;
-
-// // Font Weights
-// $font-weight--normal: 400;
-// $font-weight--bold: 700;
-
-// // Font Line Height
-// $font-line-height: 1.4;
-
-// /* Font Sizing
-//  *************/
-
-// // Set Base Font Size
-// $font-size: 1.1rem;
-
-// // Font Size Options
-// $font-size--xs: $font-size / 1.2; // 0.91rem
-// $font-size--s: $font-size; // 1.1rem
-// $font-size--m: $font-size * 1.3; // 1.43rem
-// $font-size--l: $font-size * 1.5; // 1.65rem
-// $font-size--xl: $font-size * 2.1; // 2.31rem
-
-/* Color Palette
- ***************/
-
-// // Color Palette Options
-// $color-white: #fefefe;
-// $color-black: #000013;
-// $color-indigo: #4b0082;
-// $color-indigo--lighter: #6b00b8;
-// $color-indigo--darker: #2b0049;
-
-// /* Gutter Spacing
-//  ****************/
-
-// // Set Base Gutter Spacing for Padding and Margins
-// $gutter: 1.5rem;
-
-// // Gutter Spacing Options
-// $gutter--xxxxs: $gutter / 7; // 0.214rem
-// $gutter--xxxs: $gutter / 4; // 0.375rem
-// $gutter--xxs: $gutter / 3; // 0.5rem
-// $gutter--xs: $gutter / 2; // 0.75rem
-// $gutter--s: $gutter / 1.5; // 1rem
-// $gutter--m: $gutter; // 1.5rem
-// $gutter--l: $gutter * 1.5; // 2.25rem
-// $gutter--xl: $gutter * 2; // 3rem
-// $gutter--xxl: $gutter * 3; // 4.5rem
-
-// /* Breakpoints
-//  *************/
-
-// // Set Breakpoint Options
-// $breakpoint-mobile--s: 380px;
-// $breakpoint-mobile: 500px;
-// $breakpoint-tablet: 950px;
-// $breakpoint-desktop: 1200px;
-// $breakpoint-desktop--l: 1600px;
-
-// /* Transitions
-//  *************/
-
-// // Set Background Color Transition
-// $transition-background: background-color 250ms ease-in-out;
