--- src/10_styles/_theme.scss
@@ -1,19 +1,0 @@
-// @todo: Add below to a file called global.
-
-// Set the root font-size for REMs
-:root {
-    font-size: 16px;
-}
-
-// Set Site Background Color
-$site-background-color: $theme-color-background; // @todo: Change to actual color from colors!
-
-body {
-    background-color: $site-background-color; // Set Body Background Color
-
-    @media (prefers-color-scheme: dark) {
-        background-color: $color-black;
-    }
-}
-
-// @todo create typography mixin (https://scarab.style/core/design-tokens/#font-sizes-line-heights)
