+++ src/pages/style-guide.js
@@ -6,7 +6,7 @@
 import React from "react"
 import Layout from "../04_templates/MainLayout/MainLayout"
 import Image from "../01_atoms/Image/Image"
-import Icon from "../00_design_tokens/Icon/Icon"
+import Icon from "../00_design_tokens/005_iconography/Icon/Icon"
 import InLink from "../01_atoms/InLink/InLink"
 import ExLink from "../01_atoms/ExLink/ExLink"
 import Button from "../01_atoms/Button/Button"
